Introduce stable build into recipe
